Cleaning news groups

+++ src/main/java/org/luwrain/pim/news/NewsGroups.java
@@ -25,7 +25,7 @@ import org.luwrain.pim.news.*;
 public interface NewsGroups
 {
     StoredNewsGroup[] load() throws PimException;
-    StoredNewsGroup loadById(long id) throws PimException;
+    StoredNewsGroup loadById(int id) throws PimException;
     void save(NewsGroup group) throws PimException;
     void delete(StoredNewsGroup group) throws PimException;
 }
